backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.2

# LangChain stack
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langgraph==0.2.0

# Retrieval & vectorstore
chromadb==0.5.5
sentence-transformers==3.0.1
beautifulsoup4==4.12.3
